pyproject.toml
# Зависимости и конфиг ruff, mypy

[tool.poetry]
name = "websocket-chat"
version = "0.1.0"
description = ""
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
sqlalchemy = "^2.0.0"
asyncpg = "^0.29.0"
redis = "^5.0.0"
pydantic-settings = "^2.0.0"
structlog = "^23.0.0"
prometheus-client = "^0.17.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
alembic = "^1.12.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"